fix: resolve feedback comments on the attachment

+++ src/main/java/com/example/customer_service/dtos/CustomerSaldoDTO.java
@@ -3,9 +3,11 @@ package com.example.customer_service.dtos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CustomerSaldoDTO {
     private Long id;
-    private Double saldo;
+    private BigDecimal saldo;
 }
